Add tolerance to any exception during tcp buffer processing
In ideal conditions only json exceptions are thrown but bugs
or unexpected low-level conditions can lead to other types of
exceptions when the tcp buffer does not have a complete json
string yet. This commit allows the tcp session to continue
waiting for more data in the hope the buffer will parse at some
point. If the watchdog timer kicks in the connection will be
reset anyways and the client will send an 'invalid' response.